Extract app runtime into workspace crate
