+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["crates/haqor-morphology"]
+members = ["crates/haqor-morphology", "crates/haqor-runtime"]
 resolver = "3"
 
 [package]
@@ -18,7 +18,7 @@ default = []
 # self-contained binary. Off by default: the CLI and tests read data/ off
 # disk, and the haqor app ships the files as Flutter assets — both via the
 # file-backed `Bible::open`.
-embedded = ["dep:rust-embed", "rusqlite/serialize"]
+embedded = ["haqor-runtime/embedded"]
 
 # Tests exercise the morphology generator heavily (the reverse-index test
 # generates the paradigm of every 22³ root); unoptimized they are unusably
@@ -35,9 +35,9 @@ reqwest = { version = "0.12", features = ["blocking", "json", "gzip"] }
 anyhow = "1.0.81"
 bytes = "1.10.1"
 rusqlite = { version = "0.37.0", features = ["bundled", "functions"] }
-rust-embed = { version = "8.7.2", features = ["debug-embed"], optional = true }
 quick-xml = "0.37"
 serde_json = "1.0"
 csv = "1.3"
 rayon = "1.12.0"
 haqor-morphology = { path = "crates/haqor-morphology" }
+haqor-runtime = { path = "crates/haqor-runtime" }
