+++ crates/haqor-runtime/src/vocab_gloss.rs
