+++ src/generate/harness.rs
@@ -33,12 +33,12 @@ use quick_xml::events::Event;
 use rayon::prelude::*;
 use rusqlite::Connection;
 
-use super::hebrew_db::normalize_surface;
 use super::lexicon_db::load_root_inventory;
 use super::prefilter::Prefilter;
 use crate::morphology::{
     Binyan, Form, ReverseIndex, parse_word_indexed, parse_word_indexed_disambiguated,
 };
+use haqor_runtime::normalize_surface;
 
 /// OT books are 1..=39 in Haqor numbering; NT starts at 40.
 const NT_FIRST_BOOK: u8 = 40;
